+++ LockerService.Application/Common/Services/IOrderTimeoutService.cs
@@ -2,5 +2,5 @@ namespace LockerService.Application.Common.Services;
 
 public interface IOrderTimeoutService
 {
-    Task CancelExpiredOrder(int orderId, DateTimeOffset time);
+    Task CancelExpiredOrder(long orderId, DateTimeOffset time);
 }
