+++ LockerService.Application/Dashboard/Queries/GetOrderDashboardRequest.cs
@@ -2,7 +2,7 @@ namespace LockerService.Application.Dashboard.Queries;
 
 public class GetOrderDashboardRequest : IRequest<OrderDashboardResponse>
 {
-    public int? LockerId { get; set; }
+    public long? LockerId { get; set; }
     
     public DateTimeOffset? From { get; set; }
