+++ LockerService.Application/Hardwares/Commands/RemoveHardwareCommand.cs
@@ -2,7 +2,7 @@ namespace LockerService.Application.Hardwares.Commands;
 
 public class RemoveHardwareCommand : IRequest<StatusResponse>
 {
-    public int LockerId { get; set; }
+    public long LockerId { get; set; }
     
-    public int HardwareId { get; set; }
+    public long HardwareId { get; set; }
 }
