+++ LockerService.Application/Lockers/Models/BoxStatus.cs
@@ -6,11 +6,11 @@ public class BoxStatus
     
     public bool IsAvailable { get; set; }
     
-    public int? OrderId { get; set; }
+    public long? OrderId { get; set; }
     
     public OrderStatus? OrderStatus { get; set; }
     
-    public int? ServiceId { get; set; }
+    public long? ServiceId { get; set; }
     
     public string? ServiceName { get; set; }
