+++ LockerService.Application/Lockers/Queries/GetLockerQuery.cs
@@ -2,5 +2,5 @@ namespace LockerService.Application.Lockers.Queries;
 
 public class GetLockerQuery : IRequest<LockerDetailResponse>
 {
-    public int LockerId { get; init; }
+    public long LockerId { get; init; }
 }
