+++ LockerService.Application/Staffs/Commands/AssignStaffCommand.cs
@@ -11,7 +11,11 @@ public class AssignStaffCommandValidator : AbstractValidator<AssignStaffCommand>
 
 public class AssignStaffCommand : IRequest<StatusResponse>
 {
-    [JsonIgnore] public int Id { get; set; } = default!;
-    [JsonIgnore] public int StoreId { get; set; } = default!;
-    public int LockerId { get; set; } = default!;
+    [JsonIgnore] 
+    public long Id { get; set; } = default!;
+    
+    [JsonIgnore] 
+    public long StoreId { get; set; } = default!;
+    
+    public long LockerId { get; set; } = default!;
 }
