+++ LockerService.Application/Staffs/Commands/DeactivateStaffCommand.cs
@@ -9,6 +9,10 @@ public class DeactivateStaffCommandValidator : AbstractValidator<DeactivateStaff
 
 public class DeactivateStaffCommand : IRequest<StatusResponse>
 {
-    [JsonIgnore] public int Id { get; set; } = default!;
-    [JsonIgnore] public int StoreId { get; set; } = default!;
+    [JsonIgnore] 
+    
+    public long Id { get; set; } = default!;
+    
+    [JsonIgnore] 
+    public long StoreId { get; set; } = default!;
 }
