+++ LockerService.Application/Stores/Models/StoreResponse.cs
@@ -2,10 +2,15 @@ namespace LockerService.Application.Stores.Models;
 
 public class StoreResponse
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
+    
     public string Name { get; set; } = default!;
+    
     public string? ContactPhone { get; set; }
+    
     public StoreStatus Status { get; set; }
+    
     public LocationResponse Location { get; set; } = default!;
+    
     public string? Image { get; set; }
 }
