+++ LockerService.Application/Stores/Queries/GetStoreQuery.cs
@@ -2,5 +2,5 @@ namespace LockerService.Application.Stores.Queries;
 
 public class GetStoreQuery : IRequest<StoreDetailResponse>
 {
-    public int StoreId { get; init; }
+    public long StoreId { get; init; }
 }
