+++ LockerService.Domain/Entities/Store.cs
@@ -7,28 +7,17 @@ namespace LockerService.Domain.Entities;
 [Table("Store")]
 public class Store : BaseAuditableEntity
 {
-    [Key] public int Id { get; set; }
+    [Key] 
+    public long Id { get; set; }
+    
     public string Name { get; set; } = default!;
     
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
-    
     public string? ContactPhone { get; set; }
+    
     public StoreStatus Status { get; set; } = StoreStatus.Active;
-    public int LocationId { get; set; }
+    
+    public long LocationId { get; set; }
+    
     public Location Location { get; set; } = default!;
 
     public string? Image { get; set; }
